+++ preparation/t651/authorities/ondeverte.py
@@ -1,0 +1,23 @@
+import re
+
+def read(path):
+    stations = []
+
+    with open(path) as f:
+        reading = False
+
+        for line in f:
+            line = line.strip()
+
+            match = re.match(r'^([\w+ ,/]+)\s+([0-9]{2})', line)
+
+            if match:
+                station_name = match.group(1).strip()
+                zones = [int(match.group(2))]
+                stations.append((station_name, zones, "OndeVerte"))
+
+    return stations
+
+if __name__ == "__main__":
+    import sys
+    print(read(sys.argv[1]))